+++ moveit_ros/visualization/motion_planning_rviz_plugin/src/motion_planning_frame.cpp
@@ -663,7 +663,7 @@ void MotionPlanningFrame::tabChanged(int index)
     selectedCollisionObjectChanged();
 }
 
-void MotionPlanningFrame::updateSceneMarkers(float /*wall_dt*/, float /*ros_dt*/)
+void MotionPlanningFrame::updateSceneMarkers(double /*wall_dt*/, double /*ros_dt*/)
 {
   if (scene_marker_)
     scene_marker_->update();
