+++ moveit_ros/visualization/rviz_plugin_render_tools/src/trajectory_visualization.cpp
@@ -256,7 +256,7 @@ void TrajectoryVisualization::changedShowTrail()
 
   int stepsize = trail_step_size_property_->getInt();
   // always include last trajectory point
-  trajectory_trail_.resize(static_cast<int>(std::ceil((t->getWayPointCount() + stepsize - 1) / (float)stepsize)));
+  trajectory_trail_.resize(static_cast<int>(std::ceil((t->getWayPointCount() + stepsize - 1) / (double)stepsize)));
   for (std::size_t i = 0; i < trajectory_trail_.size(); ++i)
   {
     int waypoint_i = std::min(i * stepsize, t->getWayPointCount() - 1);  // limit to last trajectory point
@@ -363,15 +363,15 @@ void TrajectoryVisualization::interruptCurrentDisplay()
     animating_path_ = false;
 }
 
-float TrajectoryVisualization::getStateDisplayTime()
+double TrajectoryVisualization::getStateDisplayTime()
 {
   constexpr char default_time_string[] = "3x";
-  constexpr float default_time_value = -3.0f;
+  constexpr double default_time_value = -3.0f;
 
   std::string tm = state_display_time_property_->getStdString();
   boost::trim(tm);
 
-  float type;
+  double type;
 
   if (tm.back() == 'x')
   {
@@ -390,7 +390,7 @@ float TrajectoryVisualization::getStateDisplayTime()
   tm.resize(tm.size() - 1);
   boost::trim_right(tm);
 
-  float value;
+  double value;
   try
   {
     value = std::stof(tm);
@@ -415,7 +415,7 @@ void TrajectoryVisualization::dropTrajectory()
   drop_displaying_trajectory_ = true;
 }
 
-void TrajectoryVisualization::update(float wall_dt, float sim_dt)
+void TrajectoryVisualization::update(double wall_dt, double sim_dt)
 {
   if (drop_displaying_trajectory_)
   {
@@ -473,7 +473,7 @@ void TrajectoryVisualization::update(float wall_dt, float sim_dt)
     {
       current_state_time_ += wall_dt;
     }
-    float tm = getStateDisplayTime();
+    double tm = getStateDisplayTime();
 
     if (trajectory_slider_panel_ && trajectory_slider_panel_->isVisible() && trajectory_slider_panel_->isPaused())
     {
@@ -488,7 +488,7 @@ void TrajectoryVisualization::update(float wall_dt, float sim_dt)
     else if (tm < 0.0)
     {
       // using realtime factors: skip to next waypoint based on elapsed display time
-      const float rt_factor = -tm;  // negative tm is the intended realtime factor
+      const double rt_factor = -tm;  // negative tm is the intended realtime factor
       while (current_state_ < waypoint_count &&
              (tm = displaying_trajectory_message_->getWayPointDurationFromPrevious(current_state_ + 1) / rt_factor) <
                  current_state_time_)
